+++ auth-old/auth/__init__.py
@@ -13,9 +13,6 @@
 'Pyramid WSGI configuration'
 from pyramid_beaker import session_factory_from_settings, set_cache_regions_from_settings
 
-from auth.views import users
-
-
     # Configure sessions and caching
     if 'session.secret' not in settings:
         settings['session.secret'] = make_random_string(32)
@@ -23,6 +20,3 @@ from auth.views import users
     set_cache_regions_from_settings(settings)
     # Configure mailer
     config.include('pyramid_mailer')
-    # Configure routes for user account management
-    config.scan(users)
-    config.include(users.add_routes)