+++ auth/CHANGES.rst
@@ -1,7 +1,9 @@
 0.0
 ---
 
-- Made tests for public, protected, privileged pages
 - Made data models
-- Made public, protected, privileged pages
+- Made public, protected, privileged views
 - Made login and logout views
+- Made functional tests for page_public, page_protected, page_privileged
+- Made functional tests for user_login and user_logout
+- Made user_index